refactor shortcodes to shortcode dir
