--- docker/bin/test-server-var-set.php
@@ -1,3 +1,0 @@
-<?php
-
-$_SERVER['HTTP_X_WPENGINE_SEGMENT'] = 'smoketest';
