+++ src/class-send-vary-header.php
@@ -13,11 +13,11 @@ if ( ! defined( 'ABSPATH' ) ) {
 }
 
 /**
- * Class SendVaryHeader
+ * Class Send_Vary_Header
  *
  * Use WordPress's send_headers hook to send a vary header if a cache segment is provided
  */
-class SendVaryHeader {
+class Send_Vary_Header {
 
 	/**
 	 * Name of header to segment on
@@ -36,21 +36,21 @@ class SendVaryHeader {
 
 		// send vary header in response when segment request header is present.
 		if ( $this->header_name ) {
-			$this->hook_wp_headers();
+			$this->add_action();
 		}
 	}
 
 	/**
 	 * Send headers via WordPress hook
 	 */
-	public function hook_wp_headers() {
-		add_action( 'send_headers', array( $this, 'send_vary_header' ) );
+	public function add_action() {
+		add_action( 'send_headers', array( $this, 'set_vary_header' ) );
 	}
 
 	/**
 	 * Set the Vary response header to allow segmented caching feature to work
 	 */
-	public function send_vary_header() {
+	public function set_vary_header() {
 		header( 'Vary: X-WPENGINE-SEGMENT' );
 	}
 }
