+++ test/smoke/test_plugin_activate.py
