+++ test/smoke/test_segment_cache_display_with_out_var.py
