+++ test/smoke/test_segment_cache_display_with_var.py
