+++ test/unit/shortcode/class-display-segment-test.php
@@ -1,87 +1,16 @@
 <?php
 /**
- * Unit tests for the Shortcode class of the Segment Cache for WP Engine plugin
+ * Unit tests for the DisplaySegment shortcode class of the Segment Cache for WP Engine plugin
  *
  * @package segment-cache-for-wp-engine
  */
 
-namespace SegmentCacheWPE;
+namespace SegmentCacheWPE\Shortcode;
 
 /**
- * Unit tests for Shortcode class
+ * Unit tests for DisplaySegment shortcode class
  */
-class Shortcode_Test extends \WP_UnitTestCase {
-
-	/**
-	 * Test validating attributes passed by humans to set_segment shortcode
-	 */
-	public function test_validate_set_segment_atts() {
-		$shortcode = new Shortcode();
-
-		// Test no attributes.
-		$atts   = array();
-		$expect = $shortcode->default_set_segment_atts;
-		$actual = $shortcode->validate_set_segment_atts( $atts );
-		$this->assertEquals( $expect, $actual );
-
-		// Test non-valid attribute.
-		$atts   = array( 'blah' => '123' );
-		$expect = $shortcode->default_set_segment_atts;
-		$actual = $shortcode->validate_set_segment_atts( $atts );
-		$this->assertEquals( $expect, $actual );
-
-		// Tests html escaping.
-		$atts        = array( 'path' => '<script></script>' );
-		$atts_expect = array( 'path' => '&lt;script&gt;&lt;/script&gt;' );
-		$expect      = array_merge( $shortcode->default_set_segment_atts, $atts_expect ); // Latter keys overwrite exiting.
-		$actual      = $shortcode->validate_set_segment_atts( $atts );
-		$this->assertEquals( $expect, $actual );
-
-		// Test replacing attributes.
-		$atts   = array(
-			'path'   => '/blog/',
-			'expire' => 12345,
-			'secure' => true,
-		);
-		$expect = array_merge( $shortcode->default_set_segment_atts, $atts ); // Latter keys overwrite exiting.
-		$actual = $shortcode->validate_set_segment_atts( $atts );
-		$this->assertEquals( $expect, $actual );
-	}
-
-	/**
-	 * Test setting cache segment cookie
-	 */
-	public function test_set_segment() {
-		$shortcode = new Shortcode();
-
-		// Test when no attributes passed.
-		$atts   = array();
-		$result = $shortcode->set_segment( $atts );
-		$this->assertNull( $result );
-
-		// Test when no segment-name attribute passed.
-		$atts   = array( 'path' => '/blah' );
-		$result = $shortcode->set_segment( $atts );
-		$this->assertNull( $result );
-
-		// Test that validate_set_segment_atts and setcookie are run when valid segment-name passed.
-		$shortcode_mock = $this->getMockBuilder( '\SegmentCacheWPE\Shortcode' )
-			->disableOriginalConstructor()
-			->setMethods( array( 'validate_set_segment_atts', 'setcookie' ) )
-			->getMock();
-
-		$shortcode_mock->expects( $this->once() )
-			->method( 'validate_set_segment_atts' )
-			->willReturn( null );
-
-		$shortcode_mock->expects( $this->once() )
-			->method( 'setcookie' )
-			->willReturn( true );
-
-		$atts   = array( 'segment-name' => 'blah' );
-		$result = $shortcode_mock->set_segment( $atts );
-		$this->assertTrue( $result );
-	}
+class DisplaySegment_Test extends \WP_UnitTestCase {
 
 	/**
 	 * Test display segmented content
@@ -90,7 +19,7 @@ class Shortcode_Test extends \WP_UnitTestCase {
 		$atts    = array();
 		$content = 'Hello, world!';
 
-		$shortcode_mock = $this->getMockBuilder( '\SegmentCacheWPE\Shortcode' )
+		$shortcode_mock = $this->getMockBuilder( '\SegmentCacheWPE\Shortcode\Display_Segment' )
 			->disableOriginalConstructor()
 			->setMethods( array( 'escape_content', 'should_show_content' ) )
 			->getMock();
@@ -114,7 +43,7 @@ class Shortcode_Test extends \WP_UnitTestCase {
 	 * Test escaping content
 	 */
 	public function test_escape_content() {
-		$shortcode       = new Shortcode();
+		$shortcode       = new Display_Segment();
 		$content         = '<b>Hello, world!</b>';
 		$escaped_content = '&lt;b&gt;Hello, world!&lt;/b&gt;';
 
@@ -148,8 +77,8 @@ class Shortcode_Test extends \WP_UnitTestCase {
 	 * Test should show content
 	 */
 	public function test_should_show_content() {
-		$shortcode             = new Shortcode();
-		$shortcode_with_header = new Shortcode( 'foo' );
+		$shortcode             = new Display_Segment();
+		$shortcode_with_header = new Display_Segment( 'foo' );
 
 		// Test when no attributes passed. No segment-name and no header = true.
 		$atts   = array();
